% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afgencomp.R
\name{countKmersCertain}
\alias{countKmersCertain}
\title{Get the Kmer Counts for certainty nucleotides (ie remove ambiguous mers)}
\usage{
countKmersCertain(genomicString, k, freq = FALSE)
}
\arguments{
\item{genomicString}{A genomic string in the character string format in R}

\item{k}{The size of the kmers which the user would like to produce a count matrix for}

\item{freq}{Return the relative proportions of kmers of that kind (default FALSE - return raw counts)}
}
\value{
A vector of kmer counts that contains counts for the kmers in lexicographical 
ordering.
}
\description{
Produces a vector of the counts of all of the kmers in a particular character
string representing a genetic sequence.  This implementation will first replace all 
non-ACGT components with a gapping place holder (-) and will then proceed to 
count all mers present which do not contain any gaps.
}
\examples{
countKmersCertain(sarscvmay[1,]$sequences,5)
}
