% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afgencomp.R
\name{encodeGenome}
\alias{encodeGenome}
\title{Encode a Single Genome}
\usage{
encodeGenome(stringGenome, dimension = "4D", strategy = "AC")
}
\arguments{
\item{stringGenome}{The genomic signal the user would like to work with}

\item{dimension}{Either '2D' or '4D' character vector, '4D' by default, this 
specifies whether to use the binary four-signal encoding originally described
in 2014 by Yin and Yau, or three-vauled two-signal encoding.}

\item{strategy}{This is only used for the 2D encoding, it determines which of 
the possible 2D encodings is used, this is determined for ensembles by 
considering the concentration of nucleotides which are Adonine and Cytosine or
Adonine and Guanine.}
}
\value{
Returns a matrix of dimension 2xSignalLength, or 4xSignalLength where 
each row is determined by the encoding strategy, for 4D encoding, each row 
represents the presence or absence of a specific nucleotide, for the 2D 
encoding, the two signals of three values -1, 0, 1 are determined according 
to the strategy choosen.
}
\description{
Encodes a single genomic signal accepted as a character string into either 
a four by signal length (4 x n), or two by signal length (2 x n) matrix 
depending on whether
the user specifies 2D or 4D, that contains the encoded genomic signal.
}
\examples{
EncodedSignal2D <- encodeGenome('ACCACTTGAAGAGAGCCCGGGAT', '4D'); 
EncodedSignal4DAC <- encodeGenome('ACCACTTGAAGAGACCCGGGAT', '2D', 'AC'); 
EncodedSignal4DAG <- encodeGenome('ACCACTTGAAGAGACCCGGGAT','2D','AG'); 
}
