% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afgencomp.R
\name{encodeGenomes}
\alias{encodeGenomes}
\title{Encode an ensemble of genomes into a numerical form}
\usage{
encodeGenomes(stringGenomes, dimension = "2D")
}
\arguments{
\item{stringGenomes}{is a list containing the genomes in a string format}

\item{dimension}{is a character string either '2D' or '4D'.}
}
\value{
A list containing matrixes of various column length, but either 2 or 
4 rows.
}
\description{
Will produce a list of either the 4D or the 2D representation of a set of 
signals and return them in a list.
}
\examples{
encodeGenomes(list('ACCCCAATTAGAGGGACTTTGGGAACCGGAGAT','GGCCCAGAGGGAAACCGGT'),
              '2D'); 
encodeGenomes(list('ACCCCAATTAGAGGGACTTTGGGAACCGGAGAT','GGCCCAGAGGGAAACCGGT'), 
              '4D'); 
}
