% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afgencomp.R
\name{getPhaseSpectraEnsemble}
\alias{getPhaseSpectraEnsemble}
\title{Get the Fourier Phase Spectra for an ensemble of encoded genomic signals}
\usage{
getPhaseSpectraEnsemble(encodedEnsemble)
}
\arguments{
\item{encodedEnsemble}{a list of encoded genomes that are produced using
one of the encoding methods programmed in this package.}
}
\value{
A list of Phase spectra for the corresponding genomic sequences.
}
\description{
This is a wrapper for the getPhaseSpectraSingle function, that allows for
the direct return of phase spectra for each of the signals in an ensemble.
}
\examples{
genStrings1 <- list('ACCAAGGATATTAGGACCC','CCCCAGGGAGATTTAGG','CCCGGGAGAGATTTAG');
encStrings1 <- encodeGenomes(genStrings1);
psStrings1 <- getPhaseSpectraEnsemble(encStrings1);
}
