% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afgencomp.R
\name{getPowerSpectraEnsemble}
\alias{getPowerSpectraEnsemble}
\title{Get the Fourier Power Spectra for an ensemble of encoded genomic signals}
\usage{
getPowerSpectraEnsemble(encodedEnsemble)
}
\arguments{
\item{encodedEnsemble}{a list of encoded genomes that are produced using
one of the encoding methods programmed in this package.}
}
\value{
A list of Power spectra for the corresponding genomic sequences.
}
\description{
This is a wrapper for the getPowerSpectraSingle function, that allows for
the direct return of power spectra for each of the signals in an ensemble.
}
\examples{
genStrings1 <- list('ACCAAGGATATTAGGACCC','CCCCAGGGAGATTTAGG','CCCGGGAGAGATTTAG');
encStrings1 <- encodeGenomes(genStrings1);
psStrings1 <- getPowerSpectraEnsemble(encStrings1);
}
