% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarscvaugaligned}
\alias{sarscvaugaligned}
\title{GISAID sequence data for 1,397 SARS-CoV-2 Genomes from August 20, 2020}
\format{
A dataframe with 1,397 rows and 2 columns: 
\describe{
 \item{Sequence}{The full sequence information in character form for the data}
 \item{Header}{The full header string extracted from the FASTA provided by GISAID}
}
}
\source{
\url{https://www.gisaid.org/}
}
\usage{
sarscvaugaligned
}
\description{
A dataset containing the the header and sequence information for 1,322 
different international sequence captures for the Single-Stranded Pluse-Sense
RNA virus called the SARS-Coronavirus 2 which is responsible for the outbreak
of the Covid19 Pandemic situation which began at the end of 2019. This set has 
been aligned using the Clustal-Omega Software.
}
\keyword{datasets}
