% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afgencomp.R
\name{tstForMultiGenomes}
\alias{tstForMultiGenomes}
\title{Test Fourier Power spectra computation with random sequences 
 
 A function for performing a test of the procedures in this package with randomly
 generated genetic strings. 
 @param numStrings the number of genomic signals to generate 
 @param avLength the average length of the genomic signals (generated according to Normal Distribution)
 @param deviation the variance of the normal distribution with which to compute the vector of lengths. 
 @examples 
 tstForMultiGenomes(100,30000,30)
 @export}
\usage{
tstForMultiGenomes(numStrings = 100, avLength, deviation)
}
\description{
Test Fourier Power spectra computation with random sequences 
 
 A function for performing a test of the procedures in this package with randomly
 generated genetic strings. 
 @param numStrings the number of genomic signals to generate 
 @param avLength the average length of the genomic signals (generated according to Normal Distribution)
 @param deviation the variance of the normal distribution with which to compute the vector of lengths. 
 @examples 
 tstForMultiGenomes(100,30000,30)
 @export
}
